feat: enhance status menu integration for user and gift card updates with dynamic status handling

+++ src/app/dashboard/settings/users/user-details/user-details.ts
@@ -1,15 +1,15 @@
 import { DatePipe } from '@angular/common';
-import { Component, computed, inject, input, signal } from '@angular/core';
+import { Component, computed, inject, input, signal, linkedSignal } from '@angular/core';
 import { Router } from '@angular/router';
-import { StatusBadgeDirective } from '../../../../directives/status-badge.directive';
 import { User } from '../../settings.interface';
 import { Location } from '@angular/common';
 import { UserDeleteModal } from './user-delete-modal/user-delete-modal';
 import { UserService } from '../../users.service';
+import { StatusMenu } from '../../../../common/status-menu/status-menu';
 
 @Component({
   selector: 'app-user-details',
-  imports: [StatusBadgeDirective, DatePipe, UserDeleteModal],
+  imports: [DatePipe, UserDeleteModal, StatusMenu],
   templateUrl: './user-details.html',
   styleUrl: './user-details.scss',
 })
@@ -21,6 +21,7 @@ export class UserDetails {
   userData = input<string>()
 
   isModalOpen = signal<boolean>(false);
+  status = linkedSignal(() => this.user()?.status);
 
   user = computed<User>(() => {
     const rawJson = this.userData();
@@ -48,10 +49,17 @@ export class UserDetails {
     this.isModalOpen.set(true)
   }
 
-  onUserDeleteEvent() {
+  onUserDeleteEvent(event: boolean) {
     this.isModalOpen.set(false);
+    if(!event) return;
     this.userService.deleteUser(this.user()!, () => {
       this.router.navigate(['/dashboard/settings']);
     });
+  }
+
+  onStatusSelectedEvent(status: string) {
+    this.userService.patchUserStatus(status, this.user()!.id, () => {
+      this.status.set(status)
+    })
   }
 }
